requirements.txt
selenium>=4.9.0
pandas>=1.5.3
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
webdriver-manager>=3.8.6
